Extracted question component, extracted test data, modified ui for questions component.
