+++ src/data.tsx
@@ -1,0 +1,18 @@
+const questions = [
+  {
+    question: 'What is 1 + 1?',
+    resolved: true,
+    resolution: '2',
+    author: { username: 'Bob Smith' },
+    group: { groupName: 'Front End' },
+  },
+  {
+    question: 'What is the meaning of life?',
+    resolved: false,
+    resolution: '',
+    author: { username: 'Angela Johnson' },
+    group: { groupName: 'Back End' },
+  },
+];
+
+export default questions;
